feature: briefcase premier jet
