+++ pyproject.toml
@@ -1,0 +1,38 @@
+[tool.briefcase]
+project_name = "MyCartable"
+bundle = "coding.cacahuete"
+version = "0.0.1"
+url = "https://github.com/jgirardet/MyCartable"
+license = "GNU General Public License v3 (GPLv3)"
+author = 'Jimmy Girardet'
+author_email = "ijkl@netc.fr"
+
+[tool.briefcase.app.mycartable]
+formal_name = "MyCartable"
+description = "My first application"
+#icon = "src/mycartable/resources/mycartable"
+sources = ['src/mycartable']
+requires = ['pyside2==5.15.0',
+            'pony==0.7.11',
+            'dickens==1.0.1',
+            'Pillow==7.1.2',
+            'mako==1.1.2',
+            'beautifulsoup4']
+
+
+[tool.briefcase.app.mycartable.macOS]
+requires = []
+
+[tool.briefcase.app.mycartable.linux]
+requires = []
+system_requires = []
+
+[tool.briefcase.app.mycartable.windows]
+requires = []
+
+# Mobile deployments
+[tool.briefcase.app.mycartable.iOS]
+requires = []
+
+[tool.briefcase.app.mycartable.android]
+requires = []
