+++ requirements.txt
@@ -1,21 +1,22 @@
 pyside2==5.15.0
 pony==0.7.11
 dickens==1.0.1
-pyinstaller==3.6
+# pyinstaller==3.6
 Pillow==7.1.2
 mako==1.1.2
+beautifulsoup4
 
 #dev
 aqtinstall==0.9
-beautifulsoup4
 black
 coverage
 freezegun
-jsbeautifier==1.11
+jsbeautifier
 mimesis
 pytest
 pytest-freezegun
 pytest-qt
 PyPDF2
 mypy
+briefcase
