+++ src/mycartable/package/database_mixins/equation_mixin.py
