+++ src/mycartable/package/exceptions.py
