+++ src/mycartable/package/operations/equation.py
